% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron.R
\name{cron2eng}
\alias{cron2eng}
\title{Generate English Interpretation of Valid Cron Strings}
\usage{
cron2eng(cron = "* * * * *")
}
\arguments{
\item{cron}{Character. A valid cron string composed of five cron expressions
separated by spaces).}
}
\value{
Character. An English sentence interpretation of the input cron
expression.
}
\description{
Demystify the meaning of a valid cron string by converting it to an
equivalent sentence in English. Can take the output from \code{\link{r2cron}}
for example. (Under development.)
}
\examples{
\dontrun{cron2eng("1,2,5 2-3 * 1/3 5")}

}
