% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialga-package.R
\docType{package}
\name{dialga-package}
\alias{dialga}
\alias{dialga-package}
\title{dialga: Build Cron Strings From R Expressions}
\description{
Provide R objects as inputs and get a standard cron expression in
    return. These strings are in a special format that's recognised by various
    systems that are designed to help you schedule computer tasks. It's easy to
    forget the required esoteric structure, hence the existence of web services
    like <https://crontab.guru/> and this package.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
