% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as_cron}
\alias{.as_cron}
\title{Internal R Expression to Cron String Procedure}
\usage{
.as_cron(x, p)
}
\arguments{
\item{x}{The user's input for a time period argument in the
\code{\link{r2cron}} function}

\item{p}{Valid cron time periods used in the \code{\link{r2cron}} function.}
}
\value{
A character string.
}
\description{
Convert time-period arguments of \code{\link{r2cron}} from R expressions to
the most appropriate cron expression format and paste them together.
}
