% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop}
\alias{.stop}
\title{Internal Stopifnot Procedure}
\usage{
.stop(x, p = c("minutes", "hours", "days_month", "months", "days_week"))
}
\arguments{
\item{x}{The user's input for a time period argument in the
\code{\link{r2cron}} function}

\item{p}{Valid cron time periods used in the \code{\link{r2cron}} function.}
}
\value{
Error if it fails.
}
\description{
Check that inputs to each time-period argument of \code{\link{r2cron}} are
valid.
}
