% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.vec2eng}
\alias{.vec2eng}
\title{Convert a Vector to a Sentence}
\usage{
.vec2eng(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A character string.
}
\description{
Paste elements of a vector into sentence form, so that items are separated by
commas and the last is separated with the word 'and'. Used for output in
\code{\link{cron2eng}}.
}
