% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron.R
\name{r2cron}
\alias{r2cron}
\title{Build A Cron String From R Expressions}
\usage{
r2cron(
  minutes = "*",
  hours = "*",
  days_month = "*",
  months = "*",
  days_week = "*",
  clip = TRUE
)
}
\arguments{
\item{minutes}{Minutes past the hour, integers 0 to 59 inclusive or asterisk.}

\item{hours}{Hours on a 24-hour clock, integers 0 to 23 inclusive or
asterisk.}

\item{days_month}{Day number of the month, integers 1 to 31 inclusive or
asterisk.}

\item{months}{Month number, integers 1 to 12 inclusive or asterisk.}

\item{days_week}{Day of the week, integers 0 to 6 inclusive (zero-indexed
starting with Sunday) or asterisk.}

\item{clip}{Logical. Copy to clipboard?}
}
\value{
A character string (copied to the clipboard at the user's
discretion).
}
\description{
Provide an R expression to each of the arguments, which are different time
periods (minutes, hours, days of the month, months, days of the week), and
receive an equivalent cron-expression string in return. See 'arguments' for
valid value ranges and 'details' for valid input formats.
}
\details{
The following input formats are acceptable for all the time period
arguments:
\itemize{
\item a single asterisk, like \code{"*"}, which is a placeholder character
meaning 'all valid integer values for the given time period' (i.e.
\code{minutes = "*"} means 'every minute')
\item a single numeric value, like \code{1}
\item an integer sequence with increments of 1, like \code{1:3}
\item an integer vector, like \code{c(2, 3, 5)}
\item an equally-spaced sequence of integers with specific start and end
values, like \code{seq(0, 59, 5)}
}
}
