% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialga-package.R
\docType{package}
\name{dialga-package}
\alias{dialga}
\alias{dialga-package}
\title{dialga: Build and Interpret Cron Strings with R}
\description{
Provide R integer vectors as inputs and get a standard cron string
    in return. You can also supply a cron string and get an English 
    interpretation in return.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/matt-dray/dialga/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
